Refactor semantics of core modules
