+++ pkg/view/flag.go
@@ -22,42 +22,16 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package format
+package view
 
-import (
-	"github.com/olekukonko/tablewriter"
-	"github.com/urfave/cli/v2"
+const (
+	FlagOutput = "output"
 )
 
-var (
-	headerColor = tablewriter.Colors{tablewriter.FgHiBlueColor}
+type OutputOption string
+
+const (
+	Table OutputOption = "table"
+	JSON  OutputOption = "json"
+	Card  OutputOption = "card"
 )
-
-func PrintTable(c *cli.Context, items []interface{}, opts *PrintOptions) {
-	fields := opts.Fields
-	table := tablewriter.NewWriter(opts.Pager)
-	table.SetBorder(false)
-	table.SetColumnSeparator(opts.Separator)
-
-	if !opts.NoHeader {
-		table.SetHeader(fields)
-		headerColors := make([]tablewriter.Colors, len(fields))
-		for i := range headerColors {
-			headerColors[i] = headerColor
-		}
-		table.SetHeaderColor(headerColors...)
-		table.SetHeaderLine(false)
-	}
-
-	rows := extractFieldValues(items, fields)
-
-	for _, row := range rows {
-		columns := make([]string, len(row))
-		for j, column := range row {
-			columns[j] = formatField(c, column)
-		}
-		table.Append(columns)
-	}
-	table.Render()
-	table.ClearRows()
-}
