+++ pkg/view/table.go
@@ -22,55 +22,42 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package format
+package view
 
 import (
-	"reflect"
-	"strings"
+	"github.com/olekukonko/tablewriter"
+	"github.com/urfave/cli/v2"
 )
 
-func extractNestedFields(field string) []string {
-	return strings.Split(field, ".") // results in ex. "Execution", "RunId"
-}
+var (
+	headerColor = tablewriter.Colors{tablewriter.FgHiBlueColor}
+)
 
-func extractFieldValues(items []interface{}, fields []string) [][]interface{} {
-	if len(fields) == 0 {
-		// dynamically examine fields
-		if len(items) == 0 {
-			return [][]interface{}{}
+func PrintTable(c *cli.Context, items []interface{}, opts *PrintOptions) {
+	fields := opts.Fields
+	table := tablewriter.NewWriter(opts.Pager)
+	table.SetBorder(false)
+	table.SetColumnSeparator(opts.Separator)
+
+	if !opts.NoHeader {
+		table.SetHeader(fields)
+		headerColors := make([]tablewriter.Colors, len(fields))
+		for i := range headerColors {
+			headerColors[i] = headerColor
 		}
-		e := reflect.ValueOf(items[0])
-		for e.Type().Kind() == reflect.Ptr {
-			e = e.Elem()
-		}
-		t := e.Type()
-		for i := 0; i < e.NumField(); i++ {
-			fields = append(fields, t.Field(i).Name)
-		}
+		table.SetHeaderColor(headerColors...)
+		table.SetHeaderLine(false)
 	}
 
-	var result = make([][]interface{}, len(items))
-	for i, item := range items {
-		result[i] = make([]interface{}, len(fields))
-		val := reflect.ValueOf(item)
-		// var columns []string
-		for j, field := range fields {
-			nestedFields := extractNestedFields(field)
-			var col interface{}
-			for _, nField := range nestedFields {
-				for val.Type().Kind() == reflect.Ptr {
-					// we want the struct value to be able to get a field by name
-					val = val.Elem()
-				}
-				val = val.FieldByName(nField)
-				col = val.Interface()
-				val = reflect.ValueOf(col)
-			}
-			result[i][j] = col
+	rows := extractFieldValues(items, fields)
 
-			val = reflect.ValueOf(item)
+	for _, row := range rows {
+		columns := make([]string, len(row))
+		for j, column := range row {
+			columns[j] = formatField(c, column)
 		}
+		table.Append(columns)
 	}
-
-	return result
+	table.Render()
+	table.ClearRows()
 }
